backend/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
numpy==1.23.3
pymongo==3.12.3
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
